Sound volume keys on chaos
Using i3 bindings, as it avoids having another program (xbindkeys) to manage.
All in one place.

Using a script for pulseaudio that tries to « change the active sink ».
It filters sinks that are running (palying sounds), and only changes these.

Fixing some script chmod.
Removing xbindkeysrc as it is redundant with i3 bindings capability.